Converted the game over to run on real units (kilograms, meters, etc) with conversion for viewing on screen, cause real scales are too big to display all at once

+++ Assets/SCRIPTS/CelestialBodies/CelestialBodyResource.cs
@@ -6,7 +6,7 @@ namespace TerraformingGame
         // represents a resource in a layer (not in transport?)
 
         public ResourceType type;
-        public float amount; // amount in cubic meters
+        public double amount; // amount in cubic meters
 
         public float meltingPoint { get { return this.type.GetMeltingPoint(); } }
         public float boilingPoint { get { return this.type.GetBoilingPoint(); } }
